feat: get legal moves

+++ lib/board_implementation/naiveCheckCheck.cpp
@@ -26,10 +26,12 @@ bool CFBoard::naiveCheckCheck(bool color, int coordA = -1, int coordB = -1) {
     uint64_t thisKingBoard = kingBoard & getColorBitBoard(color);
     uint64_t kingTile = 63ll - __builtin_clzll(thisKingBoard);
     // make sure kingTile is not coordA or coordB
-    uint64_t otherBoard =
-        getColorBitBoard(!color) & ~(1ll << coordA) & ~(1ll << coordB);
+    uint64_t otherBoard = getColorBitBoard(!color) &
+                          ~(1ll << (uint64_t)coordA) &
+                          ~(1ll << (uint64_t)coordB);
     uint64_t allBoard =
-        ((whiteBoard | blackBoard) | (1ll << coordB)) & (1ll << coordA);
+        ((whiteBoard | blackBoard) | (1ll << (uint64_t)coordB)) &
+        (1ll << (uint64_t)coordA);
     int kingRow = kingTile >> 3;
     int kingCol = kingTile & 7;
     // Check P, N
